requirements.txt
# spendee client
requests
python-dotenv
google-cloud-firestore
google-auth
jwt

# mcp
mcp[cli]
fastapi
fastmcp

# tests
pytest
pytest-asyncio
python-dotenv

# agent
mcp-agent
